Latests updates

+++ src/main/java/com/kricko/repository/AdvertSizeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kricko.domain.AdvertSize;
 
-public interface AdvertSizeRepository extends JpaRepository<AdvertSize, Integer>{
+public interface AdvertSizeRepository extends JpaRepository<AdvertSize, Long>{
 }
