+++ src/main/java/com/kricko/repository/AdvertTypeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kricko.domain.AdvertType;
 
-public interface AdvertTypeRepository extends JpaRepository<AdvertType, Integer>{
+public interface AdvertTypeRepository extends JpaRepository<AdvertType, Long>{
 }
