+++ src/main/java/com/kricko/repository/PublicationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kricko.domain.Publication;
 
-public interface PublicationRepository extends JpaRepository<Publication, Integer>{
+public interface PublicationRepository extends JpaRepository<Publication, Long>{
 }
